moved into own class and added config toggle
